+++ projects/ids-enterprise-ng/src/lib/module-nav-switcher/soho-module-nav-switcher.component.ts
@@ -36,6 +36,7 @@ export class SohoModuleNavSwitcherComponent implements AfterViewInit, AfterViewC
     generate: false,
     icon: undefined,
     changeIconOnSelect: true,
+    noSearch: false,
     moduleButtonText: undefined,
     roleDropdownLabel: undefined,
     roles: []
@@ -84,6 +85,16 @@ export class SohoModuleNavSwitcherComponent implements AfterViewInit, AfterViewC
   }
   public get changeIconOnSelect(): boolean | undefined {
     return this.modulenavswitcher?.settings.changeIconOnSelect || this._options.changeIconOnSelect;
+  }
+
+  @Input() set noSearch(val: boolean | undefined) {
+    console.log(val);
+    this._options.noSearch = val;
+    this.updated({ noSearch: this._options.noSearch });
+  }
+
+  public get noSearch(): boolean | undefined {
+    return this.modulenavswitcher?.settings.noSearch || this._options.noSearch;
   }
 
   @Input() set moduleButtonText(val: string | undefined) {
@@ -185,6 +196,9 @@ export class SohoModuleNavSwitcherComponent implements AfterViewInit, AfterViewC
     this.ngZone.runOutsideAngular(() => {
       // Initialize/store instance
       this.jQueryElement = jQuery(this.elementRef.nativeElement);
+      if (this._options.noSearch)
+        this.jQueryElement.find('select').attr('data-options', `{ noSearch: true}`);
+
       this.jQueryElement.modulenavswitcher(this._options);
       this.modulenavswitcher = this.jQueryElement.data('modulenavswitcher');